+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Provide token usage and raw model API calls for OpenAI o1-preview.
+
 ## v0.3.39 (3 October 2024)
 
 - The sample transcript will now display the target for scoring in the Score Event (for newly run evaluations).